api/requirements.txt
flask
auth0-python
python-dotenv
flask-cors
flask-restful
mysql-connector-python